Major IModel/IBakedModel refactor. Compiles and loads, but needs proper testing.
Many Models and their Baked counterparts use a 'base' or 'parent' model and delegate the methods they don't specifically override. Now these delegated models are handled in a base class. Also adds perspective handling and defaultState handling. Hopefully this fixes models properly and allows the camertransforms from the json to work as it is otherwise not making it to the baked models.